Shift manager components into separate files
