+++ src/components/PaymentMethodsManager/index.ts
@@ -1,0 +1,1 @@
+export * from "./PaymentMethodsManager"
