+++ src/components/index.ts
@@ -16,3 +16,5 @@ export * from "./AuthorizedApolloProvider"
 export * from "./FencersManager"
 export * from "./EditFencerDialog"
 export * from "./LinkAssociationPanel"
+export * from "./ConnectionsManager"
+export * from "./PaymentMethodsManager"
